+++ hydromt/model/components/config.py
@@ -1,5 +1,6 @@
 """A component to write configuration files for model simulations/kernels."""
 
+from functools import reduce
 from logging import Logger, getLogger
 from os import makedirs
 from os.path import abspath, dirname, isabs, isfile, join, splitext
@@ -28,7 +29,8 @@ class ConfigComponent(ModelComponent):
     ``ConfigComponent`` data is stored as a dictionary and can be written to a file
     in yaml or toml format. The component can be used to store model settings
     and parameters that are used in the model simulations or in the model
-    settings.
+    settings. Toml config files will be read and written using `TOMLkit <https://tomlkit.readthedocs.io/en/latest/quickstart/>`__.
+    This package will preserve the order and comments in a toml file. Note, that any comments associated with sections that are to be updated will still disappear.
     """
 
     def __init__(
@@ -163,16 +165,7 @@ class ConfigComponent(ModelComponent):
             >> {'a': 99, 'b': {'c': {'d': 24}}}
         """
         self._initialize()
-        parts = key.split(".")
-        num_parts = len(parts)
-        current = cast(Dict[str, Any], self._data)
-        for i, part in enumerate(parts):
-            if part not in current or not isinstance(current[part], dict):
-                current[part] = {}
-            if i < num_parts - 1:
-                current = current[part]
-            else:
-                current[part] = value
+        self._set_nested_dict(key, value, self._data)
 
     def get_value(self, key: str, fallback=None, abs_path: bool = False) -> Any:
         """Get a config value at key(s).
@@ -311,6 +304,10 @@ class ConfigComponent(ModelComponent):
             logger.debug("Setting model config options.")
         for k, v in data.items():
             self.set(k, v)
+
+    def _set_nested_dict(self, key: str, value: Any, data: dict):
+        keys = key.split(".")
+        reduce(lambda d, k: d.setdefault(k, {}), keys[:-1], data)[keys[-1]] = value
 
     def test_equal(self, other: ModelComponent) -> Tuple[bool, Dict[str, str]]:
         """Test if two components are equal.